create profile component
