+++ src/components/motion/nominator.js
@@ -2,46 +2,8 @@ import React from "react"
 
 import { css } from "@emotion/core"
 import styled from "@emotion/styled"
-import { politicianPicture } from "../../utils"
 
-const ProfilePic = ({ name, last_name }) => {
-  const profile = { name, lastname: last_name }
-  return <img src={politicianPicture(profile)} alt="" />
-}
-
-export { ProfilePic }
-
-const ProfileContainer = styled.li`
-  list-style: none;
-  display: flex;
-  align-items: center;
-  margin-bottom: 1.08rem;
-
-  & .profile-pic {
-    width: 40px;
-    height: 40px;
-
-    clip-path: circle(50%);
-    margin-right: 15px;
-  }
-`
-
-const profile = ({ name, last_name, party }) => {
-  return (
-    <ProfileContainer>
-      <div className="profile-pic">
-        <ProfilePic name={name} last_name={last_name} />
-      </div>
-      <div className="detail">
-        <div className="name">
-          {name} {last_name}
-        </div>
-        <div className="party">{party}</div>
-      </div>
-    </ProfileContainer>
-  )
-}
-const Profile = styled(profile)``
+import Profile from "./profile"
 
 const nominator = ({ className, motion: { purposers, seconders } }) => {
   return (
